+++ judge-sdk/src/main/java/com/dwarfeng/judge/sdk/bean/entity/JSFixedFastJsonSection.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 
 public class JSFixedFastJsonSection implements Bean {
 
-    private static final long serialVersionUID = 5130460618264200912L;
+    private static final long serialVersionUID = -2690536510602856598L;
 
     public static JSFixedFastJsonSection of(Section section) {
         if (Objects.isNull(section)) {
@@ -20,6 +20,8 @@ public class JSFixedFastJsonSection implements Bean {
                 JSFixedFastJsonLongIdKey.of(section.getKey()),
                 section.getName(),
                 section.isEnabled(),
+                section.getExpected(),
+                section.getVariance(),
                 section.getRemark()
         );
     }
@@ -33,17 +35,26 @@ public class JSFixedFastJsonSection implements Bean {
     @JSONField(name = "enabled", ordinal = 3)
     private boolean enabled;
 
-    @JSONField(name = "remark", ordinal = 4)
+    @JSONField(name = "expected", ordinal = 4)
+    private double expected;
+
+    @JSONField(name = "variance", ordinal = 5)
+    private double variance;
+
+    @JSONField(name = "remark", ordinal = 6)
     private String remark;
 
     public JSFixedFastJsonSection() {
     }
 
     public JSFixedFastJsonSection(
-            JSFixedFastJsonLongIdKey key, String name, boolean enabled, String remark) {
+            JSFixedFastJsonLongIdKey key, String name, boolean enabled, double expected, double variance,
+            String remark) {
         this.key = key;
         this.name = name;
         this.enabled = enabled;
+        this.expected = expected;
+        this.variance = variance;
         this.remark = remark;
     }
 
@@ -71,6 +82,22 @@ public class JSFixedFastJsonSection implements Bean {
         this.enabled = enabled;
     }
 
+    public double getExpected() {
+        return expected;
+    }
+
+    public void setExpected(double expected) {
+        this.expected = expected;
+    }
+
+    public double getVariance() {
+        return variance;
+    }
+
+    public void setVariance(double variance) {
+        this.variance = variance;
+    }
+
     public String getRemark() {
         return remark;
     }
@@ -85,6 +112,8 @@ public class JSFixedFastJsonSection implements Bean {
                 "key=" + key +
                 ", name='" + name + '\'' +
                 ", enabled=" + enabled +
+                ", expected=" + expected +
+                ", variance=" + variance +
                 ", remark='" + remark + '\'' +
                 '}';
     }