+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - com.dwarfeng.judge.stack.bean.entity.DriverInfo
   - com.dwarfeng.judge.stack.bean.entity.JudgerInfo
   - com.dwarfeng.judge.stack.bean.entity.Section
+- 为实体对象新增属性。
+  - com.dwarfeng.judge.stack.bean.entity.Section
 
 **注意：此版本还有大量的改动会导致该版本与低版本不兼容，所有不兼容的改动使用 (不兼容) 单独标注。**